+++ packages/fuels-programs/src/asm_instructions.rs
