+++ packages/fuels-programs/src/asm_instructions/contract_call.rs
