+++ packages/fuels-programs/src/asm_instructions/cursor.rs
