+++ packages/fuels-programs/src/asm_instructions/script_and_predicate_loader.rs
